Refactor About and Resume pages
Simplified and improved the content and styling of the About and Resume pages. Updated the About page to provide a more concise and engaging introduction. Also made changes to the Resume page to enhance the career aspirations and opportunities section. These modifications aim to create a better user experience and highlight the developer's skills and goals.

Updated Tailwind theme

Added a new retroGreen color to the Tailwind theme configuration. This color can be used to create a retro-themed design. The neonGreen color was also adjusted to a new shade. These changes expand the color options available for styling the application.

Commented out theme selection in Settings page

Temporarily disabled the theme selection feature in the Settings page. This was done to prevent any issues related to dynamic theme switching. The feature will be re-enabled in a future update once it is fully implemented and tested.